Refactor form for creating a scheduled report
